Simplified API surface.
