+++ packages/near-wallet-selector/src/services/provider/ProviderService.ts
@@ -8,11 +8,13 @@ import {
   Finality,
 } from "near-api-js/lib/providers/provider";
 import { SignedTransaction } from "near-api-js/lib/transaction";
+import { Options } from "../../interfaces/Options";
+import getConfig from "../../config";
 
 export type QueryParams = { [key in string]: unknown };
 
 export interface CallFunctionParams {
-  accountId: string;
+  accountId?: string;
   methodName: string;
   args?: object;
   finality?: Finality;
@@ -28,10 +30,14 @@ export interface ViewAccountParams {
 }
 
 class ProviderService {
+  private options: Options;
   private provider: providers.JsonRpcProvider;
 
-  constructor(url: string) {
-    this.provider = new providers.JsonRpcProvider(url);
+  constructor(options: Options) {
+    const config = getConfig(options.networkId);
+
+    this.options = options;
+    this.provider = new providers.JsonRpcProvider(config.nodeUrl);
   }
 
   private parseCodeResult<Response>(res: CodeResult): Response {
@@ -47,7 +53,7 @@ class ProviderService {
   }
 
   callFunction<Response>({
-    accountId,
+    accountId = this.options.contract.contractId,
     methodName,
     args = {},
     finality = "optimistic",
